# Models/RegistrationError.cs
/*
 * Dynamic Client Registration API
 *
 * This specification defines the APIs for a TPP to submit a Software Statement Assertion to an ASPSP for the purpose of creating OAuth clients that are registered with ASPSP. 
 *
 * OpenAPI spec version: v3.1-RC1
 * Contact: ServiceDesk@openbanking.org.uk
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class RegistrationError : IEquatable<RegistrationError>
    { 
        /// <summary>
        /// Gets or Sets Error
        /// </summary>
        [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
        public enum ErrorEnum
        {
            
            /// <summary>
            /// Enum InvalidRedirectUriEnum for invalid_redirect_uri
            /// </summary>
            [EnumMember(Value = "invalid_redirect_uri")]
            InvalidRedirectUriEnum = 1,
            
            /// <summary>
            /// Enum InvalidClientMetadataEnum for invalid_client_metadata
            /// </summary>
            [EnumMember(Value = "invalid_client_metadata")]
            InvalidClientMetadataEnum = 2,
            
            /// <summary>
            /// Enum InvalidSoftwareStatementEnum for invalid_software_statement
            /// </summary>
            [EnumMember(Value = "invalid_software_statement")]
            InvalidSoftwareStatementEnum = 3,
            
            /// <summary>
            /// Enum UnapprovedSoftwareStatementEnum for unapproved_software_statement
            /// </summary>
            [EnumMember(Value = "unapproved_software_statement")]
            UnapprovedSoftwareStatementEnum = 4
        }

        /// <summary>
        /// Gets or Sets Error
        /// </summary>
        [Required]
        [DataMember(Name="error")]
        public ErrorEnum? Error { get; set; }

        /// <summary>
        /// Gets or Sets ErrorDescription
        /// </summary>
        [DataMember(Name="error_description")]
        public string ErrorDescription { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class RegistrationError {\n");
            sb.Append("  Error: ").Append(Error).Append("\n");
            sb.Append("  ErrorDescription: ").Append(ErrorDescription).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((RegistrationError)obj);
        }

        /// <summary>
        /// Returns true if RegistrationError instances are equal
        /// </summary>
        /// <param name="other">Instance of RegistrationError to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(RegistrationError other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Error == other.Error ||
                    Error != null &&
                    Error.Equals(other.Error)
                ) && 
                (
                    ErrorDescription == other.ErrorDescription ||
                    ErrorDescription != null &&
                    ErrorDescription.Equals(other.ErrorDescription)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Error != null)
                    hashCode = hashCode * 59 + Error.GetHashCode();
                    if (ErrorDescription != null)
                    hashCode = hashCode * 59 + ErrorDescription.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(RegistrationError left, RegistrationError right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(RegistrationError left, RegistrationError right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}