# Models/SupportedAlgorithms.cs
/*
 * Dynamic Client Registration API
 *
 * This specification defines the APIs for a TPP to submit a Software Statement Assertion to an ASPSP for the purpose of creating OAuth clients that are registered with ASPSP. 
 *
 * OpenAPI spec version: v3.1-RC1
 * Contact: ServiceDesk@openbanking.org.uk
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
        /// <summary>
        /// Gets or Sets SupportedAlgorithms
        /// </summary>
        
        public enum SupportedAlgorithms
        {
            
            /// <summary>
            /// Enum RS256Enum for RS256
            /// </summary>
            
            RS256Enum = RS256,
            
            /// <summary>
            /// Enum PS256Enum for PS256
            /// </summary>
            
            PS256Enum = PS256,
            
            /// <summary>
            /// Enum ES256Enum for ES256
            /// </summary>
            
            ES256Enum = ES256
        }
}